7.4.2025
Redesigned the whole CityPartTabView so that it now includes a navigation link to a detail view, where the users can switch between different views (home, photos, 3D models) > UI pending.

Added a map view to display the location of the respective city part on a map.

Added coordinates and photo gallery items into the Model.